requirements.txt
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
email-validator==2.3.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
